server/requirements.txt
pipecat-ai[silero,whisper,websocket]
aiohttp
edge-tts
python-dotenv
